Created tile class, updated board constructor
Board now holds list of tiles. Each tile is a string for type and int
for number token attached to it.
Board constructor makes a list of strings and numbers that represent
type and number token, respectively. Then for each value in both lists
previously mentioned, creates a tile and stores it in List<tile>

Updated instance.js to grab data correctly as described above